pyproject.toml
[project]
name = "dhet_app"
version = "0.1.0"
description = "Django 5 starter with Tailwind CSS v4, Flowbite, HTMX, Alpine.js, and RBAC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "cuid2>=2.0.1",
    "dj-database-url>=3.1.0",
    "django>=5.2.10",
    "django-allauth[facebook,google,socialaccount]>=65.14.0",
    "django-role-permissions>=3.2.0",
    "django-storages>=1.14.0",
    "djangorestframework>=3.16.1",
    "djlint>=1.36.4",
    "drf-spectacular>=0.29.0",
    "pillow>=10.0.0",
    "psycopg[binary]>=3.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "ruff>=0.14.14",
    "django-simple-history>=3.10",
    "whitenoise>=6.11.0",
    "django-cookie-consent>=0.9.0",
    "django-silk>=5.3.0",
    "django-debug-toolbar>=6.2.0",
]
